requirements.txt
streamlit==1.41.1
python-dotenv==1.1.0
google-cloud-aiplatform==1.88.0
google-cloud-alloydb-connector[pg8000]==1.6.0
google-auth==2.38.0
sqlalchemy==2.0.37
pg8000==1.31.2
google-genai==1.12.1
numpy==2.2.3
pandas==2.2.3
tenacity==9.0.0
pytest==8.3.5
pytest-cov==6.0.0
